+++ research/conditional/moe_layers/continuous_moe.py
@@ -7,6 +7,7 @@ import torch
 from plotly import express as px
 
 from lizrd.core import misc, nn
+import lizrd.core.init
 from lizrd.support.logging import make_histogram
 from research.conditional.utils.misc_tools import stable_softmax_temperature, entropy
 from research.conditional.utils.layer_manager import LoggingLayer
@@ -108,19 +109,19 @@ class ContinuousMoeBaseClass(LoggingLayer):
     def init_parameters(self):
         # lin1 is parameter, one dimension for experts of size dmodel to dff/n_experts
         self.lin1 = nn.Parameter(
-            misc.get_init_weight(
+            lizrd.core.init.get_init_weight(
                 (self.dm, self.n_experts, self.expert_size), fan_in=self.dm
             )
         )
 
         self.lin2 = nn.Parameter(
-            misc.get_init_weight(
+            lizrd.core.init.get_init_weight(
                 (self.dm, self.n_experts, self.expert_size), fan_in=self.expert_size
             )
         )
         # controller: send a token of size dmodel to n_experts scalars
         self.controller = nn.Parameter(
-            misc.get_init_weight((self.dm, self.n_experts), fan_in=self.dm)
+            lizrd.core.init.get_init_weight((self.dm, self.n_experts), fan_in=self.dm)
         )
 
     def log_light(self):
