+++ research/conditional/moe_layers/ff_timed.py
@@ -1,6 +1,6 @@
 from plotly import express as px
 
-from lizrd.core import misc
+from lizrd.core import modules
 from lizrd.core.misc import resolve_activation_name
 from research.conditional.utils.layer_manager import LoggingLayer, measure_time
 
@@ -11,9 +11,9 @@ class FeedForwardTimed(LoggingLayer):
         self.dmodel = dmodel
         self.no_ff = no_ff
         self.dff = dff
-        self.logging_ff_pre_relu = misc.Linear(dmodel, dff)
+        self.logging_ff_pre_relu = modules.Linear(dmodel, dff)
         self.activation = resolve_activation_name(activation_type)
-        self.logging_ff_post_relu = misc.Linear(dff, dmodel)
+        self.logging_ff_post_relu = modules.Linear(dff, dmodel)
 
     def forward(self, x):
         with measure_time(self, "logging_ff_pre_relu"):
