+++ research/conditional/moe_layers/ffs.py
@@ -4,8 +4,10 @@ from torch.nn import functional as F
 
 from lizrd.core import misc
 from lizrd.core import nn
+import lizrd.core.init
 from lizrd.core.llm import SplitLastAxis, Transpose, MergeLastAxis
-from lizrd.core.misc import EinMix
+from lizrd.core.modules import EinMix
+import lizrd.core.modules
 from lizrd.support import ash
 from lizrd.support.profile import Timer, TimerLayer
 
@@ -22,17 +24,19 @@ class RewrittenSplitFF(nn.Module):
         self.nexperts = nexperts
         self.expertsize = expertsize
 
-        self.controller = nn.Parameter(misc.get_init_weight((dm, nexperts), fan_in=dm))
+        self.controller = nn.Parameter(
+            lizrd.core.init.get_init_weight((dm, nexperts), fan_in=dm)
+        )
         self.f1 = nn.Parameter(
-            misc.get_init_weight((dm, nexperts, expertsize), fan_in=dm)
+            lizrd.core.init.get_init_weight((dm, nexperts, expertsize), fan_in=dm)
         )
         self.f2 = nn.Parameter(
-            misc.get_init_weight(
+            lizrd.core.init.get_init_weight(
                 (nexperts, expertsize, dm), fan_in=(expertsize * nexperts / sparsity)
             )
         )
 
-        self.f1b = nn.Parameter(misc.get_init_bias((nexperts, expertsize)))
+        self.f1b = nn.Parameter(lizrd.core.init.get_init_bias((nexperts, expertsize)))
         # self.f2b = nn.Parameter(misc.get_init_bias(
         #     (nexperts, dm)))
 
@@ -171,23 +175,25 @@ class SimpleSplitFF(nn.Module):
         # assert expertsets == nexperts  # TODO: remove, it shouldn't be necessary
 
         self.controller = nn.Parameter(
-            misc.get_init_weight((dm, totalexperts), fan_in=dm)
+            lizrd.core.init.get_init_weight((dm, totalexperts), fan_in=dm)
         )
         self.cp = "d e"
         self.gp = "... t d"
         self.cout = "... t e"
         self.inner = "... e f"
 
-        self.bias = nn.Parameter(misc.get_init_bias((totalexperts, expertsize)))
+        self.bias = nn.Parameter(
+            lizrd.core.init.get_init_bias((totalexperts, expertsize))
+        )
 
         self.f1p = "d e f"
         self.f1 = nn.Parameter(
-            misc.get_init_weight((dm, totalexperts, expertsize), fan_in=dm)
+            lizrd.core.init.get_init_weight((dm, totalexperts, expertsize), fan_in=dm)
         )
 
         self.f2p = "e f d"
         self.f2 = nn.Parameter(
-            misc.get_init_weight(
+            lizrd.core.init.get_init_weight(
                 (totalexperts, expertsize, dm),
                 fan_in=(expertsize * totalexperts / sparsity),
             )
@@ -311,23 +317,25 @@ class BatchSplitFF(nn.Module):
         # assert expertsets == nexperts  # TODO: remove, it shouldn't be necessary
 
         self.controller = nn.Parameter(
-            misc.get_init_weight((dm, totalexperts), fan_in=dm)
+            lizrd.core.init.get_init_weight((dm, totalexperts), fan_in=dm)
         )
         self.cp = "d e"
         self.gp = "... t d"
         self.cout = "... t e"
         self.inner = "... e f"
 
-        self.bias = nn.Parameter(misc.get_init_bias((totalexperts, expertsize)))
+        self.bias = nn.Parameter(
+            lizrd.core.init.get_init_bias((totalexperts, expertsize))
+        )
 
         self.f1p = "d e f"
         self.f1 = nn.Parameter(
-            misc.get_init_weight((dm, totalexperts, expertsize), fan_in=dm)
+            lizrd.core.init.get_init_weight((dm, totalexperts, expertsize), fan_in=dm)
         )
 
         self.f2p = "e f d"
         self.f2 = nn.Parameter(
-            misc.get_init_weight(
+            lizrd.core.init.get_init_weight(
                 (totalexperts, expertsize, dm),
                 fan_in=(expertsize * totalexperts / sparsity),
             )
@@ -338,7 +346,9 @@ class BatchSplitFF(nn.Module):
 
         self.ogp = self.gp.replace("...", "... b")
         self.controller_loss_weight = controller_loss_weight
-        self.controller_bias = nn.Parameter(misc.get_init_bias((totalexperts,)))
+        self.controller_bias = nn.Parameter(
+            lizrd.core.init.get_init_bias((totalexperts,))
+        )
 
         self.register_full_backward_hook(BatchSplitFF.backward_hook_batch_split_ff)
         self.last_x = None
@@ -519,11 +529,13 @@ class FactoredDense(nn.Module):
         assert doutput % modules == 0
         dmodule = doutput // modules
 
-        self.gating = nn.Parameter(misc.get_init_weight((modules, dinput), fan_in=1))
-        self.projection = nn.Parameter(
-            misc.get_init_weight((dinput, dmodule), fan_in=dinput)
+        self.gating = nn.Parameter(
+            lizrd.core.init.get_init_weight((modules, dinput), fan_in=1)
         )
-        self.bias = nn.Parameter(misc.get_init_bias(doutput))
+        self.projection = nn.Parameter(
+            lizrd.core.init.get_init_weight((dinput, dmodule), fan_in=dinput)
+        )
+        self.bias = nn.Parameter(lizrd.core.init.get_init_bias(doutput))
 
     def forward(self, x):
         y = misc.einsum(
@@ -566,15 +578,15 @@ def PermutationDense(dinput):
             "verA",
             nn.Sequential(
                 SplitLastAxis(sqdi, sqdi),
-                misc.EinMix(
+                lizrd.core.modules.EinMix(
                     "... a b -> ... a c", weight_shape="a b c", a=sqdi, b=sqdi, c=sqdi
                 ),
                 Transpose(),
-                misc.EinMix(
+                lizrd.core.modules.EinMix(
                     "... a b -> ... a c", weight_shape="a b c", a=sqdi, b=sqdi, c=sqdi
                 ),
                 Transpose(),
-                misc.EinMix(
+                lizrd.core.modules.EinMix(
                     "... a b -> ... a c", weight_shape="a b c", a=sqdi, b=sqdi, c=sqdi
                 ),
                 MergeLastAxis(),
