+++ research/nonlinearities/archive/temporary_code/temp_research_bert.py
@@ -1,5 +1,5 @@
 import lizrd.core.nn as nn
-from lizrd.core.misc import EinMix
+from lizrd.core.modules import EinMix
 from lizrd.support import ash
