+++ research/nonlinearities/core/research_bert.py
@@ -2,10 +2,11 @@ from collections import OrderedDict
 
 import numpy as np
 import torch
+from lizrd.core.modules import EinMix
 
 import lizrd.core.nn as nn
 from lizrd.core.llm import LowRank
-from lizrd.core.misc import EinMix, Linear
+from lizrd.core.modules import Linear
 from lizrd.support import ash
